% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.restore_ambiguous_attrs}
\alias{.restore_ambiguous_attrs}
\title{Restore Ambiguous Match Attributes}
\usage{
.restore_ambiguous_attrs(tbl, backup)
}
\arguments{
\item{tbl}{A tibble to which attributes should be attached}

\item{backup}{A named list of ambiguous match attributes (output from
`.backup_ambiguous_attrs()`)}
}
\value{
The input tibble with ambiguous match attributes attached
}
\description{
Attaches previously backed-up ambiguous match attributes to a tibble.
}
\keyword{internal}
