% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ds4306_ag.R
\docType{data}
\name{ds43_2006_sps_class}
\alias{ds43_2006_sps_class}
\title{List of species name in tab_ds43_2006 separeted by category}
\format{
A data.frame.
}
\usage{
ds43_2006_sps_class
}
\description{
The 'ds43_2006_sps_class' includes all species separeted by genus, epithet, author,
subspecies, variety, and id (position in the
\code{tab_ds43_2006}).
}
\examples{

data(ds43_2006_sps_class)

}
\keyword{datasets}
