% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ds4306_ag.R
\docType{data}
\name{ds_043_2006_ag}
\alias{ds_043_2006_ag}
\title{Species names list from DS043-2006-AG Aprueban Categorizacion de Especies
Amenazadas de Flora Silvestre}
\format{
A tibble with the following columns:
\describe{\item{categoria}{A character vector.}
\item{accepted_name}{A character vector. The list of the accepted plant taxa names according to the Taxonomic Name Resolution Service - TNRS.}
\item{accepted_family}{A character vector. The corresponding family name of the accepted_name.}}
}
\usage{
ds_043_2006_ag
}
\description{
Species names list from DS043-2006-AG Aprueban Categorizacion de Especies
Amenazadas de Flora Silvestre
}
\examples{

data(ds_043_2006_ag)
str(ds_043_2006_ag)

}
\references{
DS043-2006-AG: Aprueban Categorizacion de Especies Amenazadas de Flora Silvestre. 13-07-2006
}
\keyword{datasets}
