% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ds043.R
\name{find_ds043}
\alias{find_ds043}
\title{find_ds043}
\usage{
find_ds043(x)
}
\arguments{
\item{x}{an atomic character string or a vector with the species names that will be verified if it endemic species}
}
\value{
a vector with information for each species names tested: a) "endemic" if the species is endemic of Perú; b)  "not endemic" if the species is not endemic of Perú; c)"not binary name" if only the genera were submitted; d) "taxon state undefined" if the submitted species name has a taxonomic status undetermined.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
The function automates the process to check if a plant species is or not endemic to Perú.
}
\examples{
# Basic usage
spp <- c("Tetramerium sagasteguianum", "Miconia alpina", "Psychopsis versteegianum")
find_ds043(spp)
}
