% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{genus_match}
\alias{genus_match}
\title{Match Genus Name}
\usage{
genus_match(df, target_df = NULL)
}
\arguments{
\item{df}{A tibble containing the genus names to be matched.}

\item{target_df}{A tibble representing the threatened species database containing the reference list of threatened species.}
}
\value{
A tibble with an additional logical column genus_match indicating whether the genus was successfully matched (`TRUE`) or not (`FALSE`).
}
\description{
This function performs a direct match of genus names against the genus names listed in the threatened species database.
}
\keyword{internal}
