% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perudiv-package.R
\docType{package}
\name{perudiv-package}
\alias{perudiv}
\alias{perudiv-package}
\title{perudiv: Peruvian protected flora and fauna}
\description{
Provides access to the lists of flora and fauna protected by different laws in Peru. The user supplies plant taxonomic names and the package returns the protected species information.
}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
