% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufijo_match.R
\name{suffix_match_species_within_genus_helper}
\alias{suffix_match_species_within_genus_helper}
\title{Suffix Match Species within Genus}
\usage{
suffix_match_species_within_genus_helper(df, target_df)
}
\arguments{
\item{df}{A tibble.}

\item{target_df}{A tibble representing the threatened species database containing the reference list of threatened species.}
}
\value{
Returns a tibble with the additional logical column
suffix_match_species_within_genus, indicating whether the specific
epithet was successfully matched within the matched genus (`TRUE`)
or not (`FALSE`).
}
\description{
Function to match the specific epithet by exchanging common suffixes
within an already matched genus in the threatened species database.
}
\keyword{internal}
