% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ds4306_ag.R
\docType{data}
\name{tab_ds43_2006}
\alias{tab_ds43_2006}
\title{List of Plant Species Name accordingly with the DS043-200-AG.}
\format{
A tibble with the following columns:
\describe{\item{id_cat}{The fixed species id of the input taxon.}
\item{input_genus}{A character vector. The input genus of the corresponding species name.}
\item{input_epitheton}{A character vector. The specific epithet of the corresponding species name.}
\item{rank}{A character vector. The taxonomic rank: "species","subspecies", "variety", of the corresponding species name.}
\item{input_subspecies_epitheton}{A character vector. If the indicated rank is below species, the subspecies epithet input of the corresponding species name.}
\item{taxonomic_status}{A character vector. description if a taxon is classified as ‘accepted’, ‘synonym’, ‘no opinion’. According to the Taxonomic Name Resolution Service - TNRS.}
\item{accepted_name}{A character vector. The list of the accepted plant taxa names according to the Taxonomic Name Resolution Service - TNRS.}
\item{accepted_family}{A character vector. The corresponding family name of the accepted_name.}
\item{accepted_name_author}{A character vector. The corresponding author name of the accepted_name, staying empty if the taxonomic_status is "Synonym" or "No opinion".}}
}
\usage{
tab_ds43_2006
}
\description{
The 'tab_ds43_2006' contains records belonging to all the species DS043-200-AG.
}
\examples{

data(tab_ds43_2006)
str(tab_ds43_2006)

}
\references{
DS043-2006-AG: Aprueban Categorizacion de Especies Amenazadas de Flora Silvestre. 13-07-2006
}
\keyword{datasets}
