% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ds4306_ag.R
\docType{data}
\name{tab_ds43_2006_position}
\alias{tab_ds43_2006_position}
\title{List of the number positions of the first 3 letters of the species name in
the tab_ds43_2006}
\format{
A data frame with 305 observations on the following 3 variables.
\describe{
\item{position}{A character vector. It is the position of the first 3 letters of the species name in the tab_ds43_2006.}
\item{triphthong}{A character vector. First 3 letters of the species name in the tab_ds43_2006.}
\item{genus}{A character vector. Corresponding Genus name.} }
}
\usage{
tab_ds43_2006_position
}
\description{
The 'tab_ds43_2006_position' reports the
position (in term of number of rows) of the first three letters (triphthong)
for the plant names stored in the variable 'accepted_name' of the table
'tab_ds43_2006'. This indexing system speeds up of the search on the
largest list using the package.
}
\examples{

data(tab_ds43_2006_position)
str(tab_ds43_2006_position)
}
\keyword{datasets}
