% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_043_2006_ag.R
\docType{data}
\name{ds_043_2006_ag}
\alias{ds_043_2006_ag}
\title{Species names list from DS043-2006-AG Aprueban Categorizacion de Especies
Amenazadas de Flora Silvestre}
\format{
A tibble with 777 rows:
}
\source{
{Originally published in: DS043-2006-AG: Aprueban Categorizacion de Especies
Amenazadas de Flora Silvestre. 13-07-2006}
}
\usage{
ds_043_2006_ag
}
\description{
Species names list from DS043-2006-AG Aprueban Categorizacion de Especies
Amenazadas de Flora Silvestre
}
\keyword{internal}
